refactor: 커스텀 훅 분리 [TRL-200]
